Create Bis
The Bissection Method is a mathematical 
tool to find roots of equations. 
So , if you have an idea of the interval
 where the root can be, you set this interval on 
this code and voi'lá ! The algoritm 
returns a very accurate aproximation of the root. 
he bissection method is very intuitive and consists
 in make smaller division each time till the limit given.
 The Bolzano theorem afirms that,
 if f(a)> 0 and f(b) < 0 for any (a,b), 
then exists a c such as f(c) =0. The way 
that bissection works is to find change on 
sinal between each interval, and goes on and on updating 
the intervals according to the change of sinal.